Lucene.Net.Queries refactor: Changed sbyte > byte for CLS compliance

+++ src/Lucene.Net.Queries/Function/ValueSources/ByteFieldSource.cs
@@ -69,9 +69,9 @@ namespace Lucene.Net.Queries.Function.ValueSources
                 this.arr = arr;
             }
 
-            public override sbyte ByteVal(int doc)
+            public override byte ByteVal(int doc)
             {
-                return arr.Get(doc);
+                return (byte)arr.Get(doc);
             }
 
             public override short ShortVal(int doc)
